chore: fixing tests and linting
